Add queuable job so indexing can happen async. Also move repository functionality into trait so users have option of either extending repo or using as trait. When using as trait a few abstract methods will need to be implemented. Also we reply a little on app() but I dont see that as a huge issue, considering this is a Laravel feature anyway.
