+++ src/Indexable.php
@@ -57,4 +57,14 @@ trait Indexable
     {
         $this->classMetaData = $classMetadata;
     }
+
+    /**
+     * Get the queue connection that should be used when syncing.
+     *
+     * @return string
+     */
+    public function syncWithSearchUsing()
+    {
+        return config('queue.default');
+    }
 }
